좋습니당

+++ Location.cpp
@@ -1,16 +1,13 @@
 #include "Location.hpp"
+#include "ConfigParser.hpp"
 #include "Server.hpp"
-#include "config.hpp"
 #include <sstream>
 
 Location::Location()
-    : root("/"), allowMethod(0), autoIndex(false), clientBodySize(10240)
-{
-}
+    : root("/"), allowMethod(0), autoIndex(false), clientBodySize(10240) {}
 
-Location::~Location()
-{
-}
+Location::~Location() {}
+
 bool Location::fillLocationBlock(std::map<std::string, std::string>& mapSentence)
 {
     std::map<std::string, std::string>::iterator it = mapSentence.begin();
@@ -117,7 +114,8 @@ bool Location::setAllowMethod(std::string& sentence)
     sstream << sentence;
     while (sstream >> token)
         tokens.push_back(token);
-    for (int i = 1; i < tokens.size(); i++)
+    for (size_t i = 1; i < tokens.size(); i++)
+
     {
         if (tokens[i] == "GET")
             allowMethod |= GET;
